Refactor, implement recipe selection in meal planner edit
