+++ app/src/main/java/com/javainiai/chefskiss/data/datasources/SelectedRecipeDataSource.kt
@@ -1,0 +1,17 @@
+package com.javainiai.chefskiss.data.datasources
+
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.update
+
+class SelectedRecipeDataSource {
+
+    private val _recipeId = MutableStateFlow(null as Long?)
+
+    val recipeId = _recipeId.asStateFlow()
+
+    fun updateRecipeId(recipeId: Long) {
+        _recipeId.update { recipeId }
+    }
+
+}
