+++ app/src/main/java/com/javainiai/chefskiss/ui/components/FilterComponents.kt
@@ -1,4 +1,4 @@
-package com.javainiai.chefskiss.ui.homescreen
+package com.javainiai.chefskiss.ui.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
