+++ app/src/main/java/com/javainiai/chefskiss/ui/selectionscreen/SelectionScreenViewModel.kt
@@ -1,0 +1,11 @@
+package com.javainiai.chefskiss.ui.selectionscreen
+
+import androidx.lifecycle.ViewModel
+import com.javainiai.chefskiss.data.datasources.SelectedRecipeDataSource
+
+class SelectionScreenViewModel(private val selectedRecipeDataSource: SelectedRecipeDataSource) :
+    ViewModel() {
+    fun setSelectedRecipe(recipeId: Long) {
+        selectedRecipeDataSource.updateRecipeId(recipeId)
+    }
+}
